+++ NEWS.md
@@ -1,6 +1,7 @@
-# djprshiny 0.0.3.900
+# djprshiny 0.0.3.90*0*
 * reactives can now be passed to plot functions through ... in plot_server
 * set font to Arial before downloading chart with djpr_plot_server
+* x_expand_mult argument added to djpr_ts_linechart()
 
 # djprshiny 0.0.2.900
 * modifications to plot_server to enable nested modules